+++ src/vmware/CHANGELOG.md
@@ -1,9 +1,13 @@
 # Release History
 
-## 2.1.0 (TBD)
+## 3.0.0 (2021-06)
+This version has **breaking changes**:
+- `az vmware datastore create` has been removed. Please use `az vmware datastore net-app-volume create` or `az vmware datastore disk-pool-volume create` instead.
+
+Other changes:
 - update to AVS 2021-06-01 API
-- add az vmware private-cloud rotate-vcenter-password
-- add az vmware private-cloud rotate-nsxt-password
+- add `az vmware private-cloud rotate-vcenter-password`
+- add `az vmware private-cloud rotate-nsxt-password`
 
 ## 2.0.1 (2021-02)
 - Update the minimum az cli version to 2.11.0 [#3045](https://github.com/Azure/azure-cli-extensions/issues/3045)